chore(CI): add configuration file
- Extract infiniteScroll logic from homepage
into a resuable hook

Refs: #9
