+++ .vscode/settings.json
@@ -1,3 +1,10 @@
 {
-  "cSpell.words": ["chakra", "fontsource", "olist", "prefetch", "sellit"]
+  "cSpell.words": [
+    "chakra",
+    "circleci",
+    "fontsource",
+    "olist",
+    "prefetch",
+    "sellit"
+  ]
 }
