+++ pages/index.tsx
@@ -3,33 +3,10 @@ import { Box, Heading } from "@chakra-ui/react";
 import DataTable from "components/DataTable/DataTable";
 import useAlbumListingFetch from "hooks/useAlbumListingFetch";
 import React from "react";
-
-const useInfiniteScroll = (): {
-  limit: number;
-  handleInfiniteScroll: () => void;
-  handleLimitChange: (newLimit: number) => void;
-} => {
-  // const [page, setPage] = React.useState(1);
-  const [limit, setLimit] = React.useState<number>(20);
-
-  // Will setLimit and trigger a new fetch, Thus infinite scroll will be simulated
-  const handleInfiniteScroll = () => {
-    setLimit((prevLimit) => prevLimit + 20);
-  };
-
-  const handleLimitChange = (newLimit: number): void => {
-    setLimit(newLimit);
-  };
-
-  return {
-    limit,
-    handleInfiniteScroll,
-    handleLimitChange,
-  };
-};
+import useTableInfiniteScroll from "hooks/useTableInfiniteScroll";
 
 const HomePage: NextPage = () => {
-  const { limit, handleInfiniteScroll } = useInfiniteScroll();
+  const { limit, handleInfiniteScroll } = useTableInfiniteScroll();
 
   const { albums, handleFetch } = useAlbumListingFetch();
